separated styled components, contents, ... moved components and its relatives to more comprehend folders
