+++ src/components/contents/MoreContents.tsx
@@ -1,26 +1,26 @@
-import { MoreContentsWrapper } from './MoreContentsWrapper.styled';
+import { IContent } from '../../lib/interfaces/IContent';
 import { Tile } from '../tiles/Tile';
-import { MoreContentsTitle } from './MoreContentsTitle';
 import { MoreContentsGridWrapper } from './MoreContentsGridWrapper.styled';
-import { IContent } from '../../interfaces/IContent';
+import { MoreContentsTitle } from './MoreContentsTitle';
+import { MoreContentsWrapper } from './MoreContentsWrapper.styled';
 
 interface IContentsData<T extends IContent> {
   data: Array<T>;
   canReplace?: boolean;
 }
 
-export const MoreContents = (props: IContentsData<IContent>) => {
+export const MoreContents = ({ data, canReplace }: IContentsData<IContent>) => {
   return (
     <MoreContentsWrapper>
       <MoreContentsTitle />
       <MoreContentsGridWrapper>
-        {props.data.map((c, i) => (
+        {data.map((c, i) => (
           <Tile
             key={i.toString()}
             link={c.link}
             title={c.title}
             prefix={'detail'}
-            canReplace={props.canReplace}
+            canReplace={canReplace}
           />
         ))}
       </MoreContentsGridWrapper>
