+++ src/components/detail/description/Description.tsx
@@ -1,19 +1,10 @@
-import { DescriptionWrapper } from './DescriptionWrapper.styled';
-import styled from 'styled-components';
+import { IDescription } from '../../../lib/interfaces/IDescription';
+import { DescriptionContent, DescriptionWrapper } from './Description.styled';
 
-interface IDescription {
-  description: string;
-}
-
-const DescriptionContent = styled.article`
-  flex: 1;
-  text-align: center;
-`;
-
-export const Description = (props: IDescription) => {
+export const Description = ({ description }: IDescription) => {
   return (
     <DescriptionWrapper>
-      <DescriptionContent>{props.description}</DescriptionContent>
+      <DescriptionContent>{description}</DescriptionContent>
       <DescriptionContent>
         <p>?</p>
         <p>Mapa</p>
