+++ src/components/header/Header.content.ts
@@ -1,0 +1,9 @@
+import { IHeaderItem } from '../../lib/interfaces/IHeaderItem';
+
+export const HeaderItems: Array<IHeaderItem> = [
+  { title: 'UTB.cz', link: 'https://www.utb.cz' },
+  {
+    title: 'Aplikace UTB',
+    link: 'https://play.google.com/store/apps/details?id=cz.unizone.utb&hl=cs&gl=US',
+  },
+];
