+++ src/components/header/Header.styled.tsx
@@ -1,0 +1,26 @@
+import styled from 'styled-components';
+
+export const HeaderWrapper = styled.header`
+  height: 60px;
+  width: 100vw;
+  background-color: white;
+  display: flex;
+  align-items: center;
+`;
+
+export const HeaderItem = styled.a`
+  text-decoration: none;
+  margin-right: 60px;
+  color: black;
+`;
+
+export const AppLogoImage = styled.img`
+  margin-right: 30px;
+`;
+
+export const AppLogo = styled.div`
+  margin-right: auto;
+  margin-left: 60px;
+  display: flex;
+  cursor: pointer;
+`;
