+++ src/components/main/presets/PresetsSection.tsx
@@ -1,13 +1,14 @@
-import { PresetsSectionWrapper } from './PresetsSectionWrapper.styled';
-import { PresetsWrapper } from './PresetsWrapper.styled';
-import { PresetsMore } from './more/PresetsMore';
-import { Tile } from '../../tiles/Tile';
-import { PresetsHeader } from './header/PresetsHeader';
 import { useContext } from 'react';
 import { SectionsContext } from '../../../contexts/SectionsContext';
+import { Tile } from '../../tiles/Tile';
+import { PresetsHeader } from './header/PresetsHeader';
+import { PresetsMore } from './more/PresetsMore';
+import { PresetsSectionWrapper } from './PresetsSectionWrapper.styled';
+import { PresetsWrapper } from './PresetsWrapper.styled';
 
 export const PresetsSection = () => {
   const sections = useContext(SectionsContext);
+
   return (
     <PresetsSectionWrapper>
       <PresetsHeader />
