+++ src/components/main/presets/PresetsWrapper.styled.tsx
@@ -5,6 +5,5 @@ export const PresetsWrapper = styled.div`
   flex-wrap: wrap;
   justify-content: center;
   gap: 21px;
-  margin-left: 100px;
-  margin-right: 100px;
+  margin: 0 100px;
 `;
