+++ src/components/main/presets/header/PresetsHeader.styled.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const HeaderWrapper = styled.div`
+  margin-top: 48px;
+  margin-bottom: 48px;
+  text-align: center;
+`;
+
+export const StyledParagraph = styled.p`
+  font-style: normal;
+  font-weight: 700;
+  font-size: 36px;
+  line-height: 108%;
+`;
