+++ src/components/main/presets/header/PresetsHeader.tsx
@@ -1,21 +1,8 @@
-import styled from 'styled-components';
-
-const HeaderWrapper = styled.div`
-  margin-top: 48px;
-  margin-bottom: 48px;
-  text-align: center;
-`;
-
-const P = styled.p`
-  font-style: normal;
-  font-weight: 700;
-  font-size: 36px;
-  line-height: 108%;
-`;
+import { HeaderWrapper, StyledParagraph } from './PresetsHeader.styled';
 
 export const PresetsHeader = () => (
   <HeaderWrapper>
-    <P>Nevíš kudy kam? To je v pořádku.</P>
-    <P>Přečti si vše důležité o UTB zde.</P>
+    <StyledParagraph>Nevíš kudy kam? To je v pořádku.</StyledParagraph>
+    <StyledParagraph>Přečti si vše důležité o UTB zde.</StyledParagraph>
   </HeaderWrapper>
 );
