+++ src/components/main/rector/RectorSectionWrapper.styled.tsx
@@ -4,6 +4,5 @@ export const RectorSectionWrapper = styled.section`
   width: 100vw;
   display: flex;
   column-gap: 82px;
-  padding-top: 48px;
-  padding-bottom: 48px;
+  padding: 48px 0;
 `;
