+++ src/components/socials/Socials.styled.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const SocialsWrapper = styled.div`
+  display: flex;
+  margin-right: 60px;
+  align-items: center;
+`;
+
+export const SocialImage = styled.img`
+  margin-right: 30px;
+`;
