+++ src/components/splash/Splash.tsx
@@ -1,15 +1,15 @@
-import { SplashWrapper } from './SplashWrapper.styled';
 import { SplashH6 } from './SplashH6.styled';
+import { SplashWrapper } from './SplashWrapper.styled';
 
 interface ISplash {
   sectionBgColor: string;
   title: string;
 }
 
-export const Splash = (props: ISplash) => {
+export const Splash = ({ sectionBgColor, title }: ISplash) => {
   return (
-    <SplashWrapper backgroundColor={props.sectionBgColor}>
-      <SplashH6>{props.title}</SplashH6>
+    <SplashWrapper backgroundColor={sectionBgColor}>
+      <SplashH6>{title}</SplashH6>
     </SplashWrapper>
   );
 };
