+++ src/lib/enums/SocialTypes.ts
@@ -1,0 +1,5 @@
+export enum SocialTypes {
+  instagram,
+  facebook,
+  youtube,
+}
