+++ src/lib/interfaces/IContent.ts
