+++ src/lib/interfaces/IDescription.ts
@@ -1,0 +1,3 @@
+export interface IDescription {
+  description: string;
+}
