+++ src/lib/interfaces/IHeaderItem.ts
@@ -1,0 +1,4 @@
+export interface IHeaderItem {
+  title: string;
+  link: string;
+}
