+++ src/lib/interfaces/ISocialItem.ts
@@ -1,0 +1,7 @@
+import { SocialTypes } from '../enums/SocialTypes';
+
+export interface ISocialItem {
+  type: SocialTypes;
+  link: string;
+  alt: string;
+}
